26/11/2021: hoan thanh 95% project

+++ src/com/controller/AccountListItemController.java
@@ -4,10 +4,14 @@ import com.helper.AccountDatabaseHelper;
 import com.model.Account;
 import com.view.Navigator;
 import java.io.IOException;
+import java.util.Optional;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
@@ -71,7 +75,13 @@ public class AccountListItemController {
 
   @FXML
   private void deleteAccount () throws IOException {
-    AccountDatabaseHelper.deleteAccount(account.getId());
+    Alert alert = new Alert(AlertType.CONFIRMATION);
+    alert.setContentText("Are you sure you want to do it?");
+
+    Optional<ButtonType> option = alert.showAndWait();
+    if (option.get() == ButtonType.OK) {
+      AccountDatabaseHelper.deleteAccount(account.getId());
+    }
     Navigator.getInstance().goToAccountList();
   }
 }
