+++ src/com/controller/CategoryListItemController.java
@@ -1,10 +1,15 @@
 package com.controller;
 
+import com.helper.AccountDatabaseHelper;
 import com.helper.CategoryDatabaseHelper;
 import com.model.Category;
 import com.view.Navigator;
 import java.io.IOException;
+import java.util.Optional;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
@@ -42,8 +47,14 @@ public class CategoryListItemController {
 
   @FXML
   void deleteCategory(MouseEvent event) throws IOException {
-    CategoryDatabaseHelper.deleteCategory(category.getId());
-    Navigator.getInstance().goToCategoryList();
+    Alert alert = new Alert(AlertType.CONFIRMATION);
+    alert.setContentText("Are you sure you want to do it?");
+
+    Optional<ButtonType> option = alert.showAndWait();
+    if (option.get() == ButtonType.OK) {
+      CategoryDatabaseHelper.deleteCategory(category.getId());
+      Navigator.getInstance().goToCategoryList();
+    }
   }
 
   @FXML
